% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{ComDimFD}
\alias{ComDimFD}
\title{Federated ComDim}
\usage{
ComDimFD(loginFD, logins, variables, TOL = 1e-10)
}
\arguments{
\item{logins}{Login info}

\item{variables}{Variables}

\item{TOL}{Tolerance of 0}
}
\description{
Federated ComDim
}
