% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_aggregate.R
\name{crossAssign}
\alias{crossAssign}
\title{Cross assign

Cross assign # rather on client side}
\usage{
crossAssign(opal, symbol, value, wait = F, async = T)
}
\arguments{
\item{opal}{An opal object or list of opal objects.}

\item{symbol}{Name of an R symbol.}

\item{value}{An encoded expression with allowed assign function calls.}

\item{wait}{See opal::datashield.aggreate options. Default: FALSE.}

\item{async}{See opal::datashield.aggreate options. Default: TRUE.}
}
\description{
Cross assign

Cross assign # rather on client side
}
