% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{.federateSSCP}
\alias{.federateSSCP}
\title{Federated SSCP}
\usage{
.federateSSCP(
  loginFD,
  logins,
  funcPreProc,
  querytables,
  byColumn = TRUE,
  scale = FALSE,
  chunk = 500L,
  mc.cores = 1,
  TOL = .Machine$double.eps,
  width.cutoff = 500L,
  connRes = FALSE
)
}
\arguments{
\item{loginFD}{Login information of the FD server}

\item{logins}{Login information of data repositories}

\item{funcPreProc}{Definition of a function for preparation of raw data
matrices. Two arguments are required: conns (list of DSConnection-classes), 
symbol (name of the R symbol) (see datashield.assign).}

\item{querytables}{Vector of names of the R symbols to assign in the
DataSHIELD R session on each server in \code{logins}.
The assigned R variables will be used as the input raw data.
Other assigned R variables in \code{funcPreProc} are ignored.}

\item{byColumn}{A logical value indicating whether the input data is
centered by column or row. Default, TRUE, centering by column, with constant
variables across samples are removed. If FALSE, centering and scaling by
row, with constant samples across variables are removed.}

\item{scale}{A logical value indicating whether the variables should be
scaled to have unit variance. Default, FALSE.}

\item{chunk}{Size of chunks into what the resulting matrix is partitioned.
Default, 500L.}

\item{mc.cores}{Number of cores for parallel computing. Default, 1.}

\item{TOL}{Tolerance of 0. Default, \code{.Machine$double.eps}.}

\item{width.cutoff}{Default, 500L. See \code{deparse1}.}

\item{connRes}{A logical value indicating if the connection to \code{logins}
is returned. Default, FALSE, connections are closed.}
}
\description{
Function for computing the federated SSCP matrix.
}
\keyword{internal}
