% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{.rebuildMatrix}
\alias{.rebuildMatrix}
\title{Symmetric matrix reconstruction from bigmemomy blocks}
\usage{
.rebuildMatrix(dscblocks, mc.cores = 1)
}
\arguments{
\item{dscblocks}{List of list of bigmemory objects pointed to matrix blocks}

\item{mc.cores}{Number of cores for parallel computing}
}
\value{
The complete symmetric matrix
}
\description{
Rebuild a symmetric matrix from its partition in bigmemory objects
}
\keyword{internal}
