% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{.rebuildMatrixDscrm}
\alias{.rebuildMatrixDscrm}
\title{Symmetric matrix reconstruction}
\usage{
.rebuildMatrixDscrm(dscblocks, mc.cores = 1)
}
\arguments{
\item{dscblocks}{List of lists of bigmemory objects pointed to matrix blocks}

\item{mc.cores}{Number of cores for parallel computing. Default: 1}
}
\value{
The complete symmetric matrix
}
\description{
Rebuild a symmetric matrix from its partition bigmemory objects
}
\keyword{internal}
