% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{.rebuildMatrixrm}
\alias{.rebuildMatrixrm}
\title{Symmetric matrix reconstruction}
\usage{
.rebuildMatrixrm(matblocks, mc.cores = 1)
}
\arguments{
\item{matblocks}{List of lists of matrix blocks, obtained from
.partitionMatrix.}

\item{mc.cores}{Number of cores for parallel computing. Default, 1.}
}
\value{
The complete symmetric matrix
}
\description{
Rebuild a matrix from its partition.
}
\keyword{internal}
