% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{.toEuclidean}
\alias{.toEuclidean}
\title{Euclidean distance}
\usage{
.toEuclidean(XXt)
}
\arguments{
\item{XXt}{An SSCP matrix XX'}
}
\value{
Euclidean distance
}
\description{
Transform XX' matrix into Euclidean between samples (rows) in X
}
\keyword{internal}
