% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_assign.R
\name{dotprod}
\alias{dotprod}
\title{Dot product

Dot product of two matrices}
\usage{
dotprod(x, y)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{y}{A numeric matrix or data frame.}
}
\value{
xy
}
\description{
Dot product

Dot product of two matrices
}
