% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federateScudo.R
\name{federateScudo}
\alias{federateScudo}
\title{Federated Scudo}
\usage{
federateScudo(X,group,H=2,scale="none",option="none",threshold=1e-10)
}
\arguments{
\item{loginFD}{Login information of the FD server}

\item{logins}{Login information of the servers containing cohort data}

\item{variables}{Variables}

\item{XX}{:            list of dataframes XX = X %*% t(X)}

\item{group}{:       named list of variables for each table}
}
\value{
XX
}
\description{
Function for ComDim federated analysis on the virtual cohort combining multiple cohorts
Finding common dimensions in multitable data (Xk, k=1...K)
}
