% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{federateUMAP}
\alias{federateUMAP}
\title{Federated UMAP}
\usage{
federateUMAP(loginFD,
                    logins,
                    func,
                    symbol,
                    metric = 'euclidean',
                    chunk = 500,
                    mc.cores = 1,
                    width.cutoff = 500L,
                    ...)
}
\arguments{
\item{loginFD}{Login information of the FD server}

\item{logins}{Login information of data repositories}

\item{func}{Encoded definition of a function for preparation of raw data
matrices. Two arguments are required: conns (list of DSConnection-classes), 
symbol (name of the R symbol) (see datashield.assign).}

\item{symbol}{Encoded vector of names of the R symbols to assign in the
DataSHIELD R session on each server in \code{logins}.
The assigned R variables will be used as the input raw data.
Other assigned R variables in \code{func} are ignored.}

\item{metric}{Either \code{euclidean} or \code{correlation} for distance
metric between samples. 
For Euclidean distance, the data from each cohort will be centered (not
scaled) for each variable.
For correlation-based distance, the data from each cohort will be centered
scaled for each sample.}

\item{chunk}{Size of chunks into what the resulting matrix is partitioned.
Default, 500.}

\item{mc.cores}{Number of cores for parallel computing. Default, 1.}

\item{...}{see \code{uwot::umap}}
}
\value{
A matrix of optimized coordinates.
}
\description{
Function for UMAP federated analysis on the virtual cohort
combining multiple cohorts
}
