% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{matrix2DscFDrm}
\alias{matrix2DscFDrm}
\title{Bigmemory description of a matrix}
\usage{
matrix2DscFDrm(value)
}
\arguments{
\item{value}{Encoded value of a matrix.}
}
\value{
Bigmemory description of the given matrix
}
\description{
Bigmemory description of a matrix.
}
\keyword{internal}
