% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{pushSingMatrix}
\alias{pushSingMatrix}
\title{Push a symmetric matrix}
\usage{
pushSingMatrix(value)
}
\arguments{
\item{value}{An encoded value to be pushed}
}
\value{
Description of the pushed value

Description of the pushed value
}
\description{
Push symmetric matrix data into the federated server

Push one-column matrix data into the federated server
}
