% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{pushSymmMatrixClient}
\alias{pushSymmMatrixClient}
\title{Push a symmetric matrix}
\usage{
pushSymmMatrixClient(value)
}
\arguments{
\item{value}{An encoded value to be pushed}
}
\value{
Description of the pushed value
}
\description{
Push symmetric matrix data into the federated server
}
