% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{solveSSCP}
\alias{solveSSCP}
\title{Find X from XX' and X'X}
\usage{
solveSSCP(XXt, XtX, r, Xr, TOL = 1e-10)
}
\arguments{
\item{XXt}{XX'}

\item{XtX}{X'X}

\item{r}{A non-null vector of length \code{ncol(XtX)}}

\item{Xr}{A vector of length \code{nrow(XXt)}, equals to the product X %*% r}
}
\value{
X
}
\description{
Find X from XX' and X'X
}
\keyword{internal}
