% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{testSSCP}
\alias{testSSCP}
\title{Test federateSSCP}
\usage{
testSSCP(
  loginFD,
  logins,
  func,
  symbol,
  metric = "euclidean",
  chunk = 500,
  mc.cores = 1,
  TOL = 1e-10,
  width.cutoff = 500L,
  ...
)
}
\arguments{
\item{loginFD}{Login information of the FD server}

\item{logins}{Login information of data repositories}

\item{func}{Encoded definition of a function for preparation of raw data
matrices. 
Two arguments are required: conns (list of DSConnection-classes), 
symbol (name of the R symbol) (see datashield.assign).}

\item{symbol}{Encoded vector of names of the R symbols to assign in the
DataSHIELD R session on each server in \code{logins}.
The assigned R variables will be used as the input raw data.
Other assigned R variables in \code{func} are ignored.}

\item{chunk}{Size of chunks into what the resulting matrix is partitioned. Default: 500}

\item{mc.cores}{Number of cores for parallel computing. Default: 1.}

\item{TOL}{Tolerance of 0}

\item{byColumn}{A logical value indicating whether the input data is centered by column or row.
Default, TRUE, centering by column. Constant variables across samples are removed. 
If FALSE, centering and scaling by row. Constant samples across variables are removed.}
}
\description{
Deprecated
}
