% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{.federateSSCP}
\alias{.federateSSCP}
\title{Federated SSCP}
\usage{
.federateSSCP(
  loginFD,
  logins,
  funcPreProc,
  querytables,
  byColumn = TRUE,
  scale = FALSE,
  chunk = 500,
  mc.cores = 1,
  TOL = 1e-10
)
}
\arguments{
\item{loginFD}{Login information of the FD server}

\item{logins}{Login information of data repositories}

\item{funcPreProc}{Definition of a function for preparation of raw data matrices. 
Two arguments are required: conns (list of DSConnection-classes), 
symbol (name of the R symbol) (see datashield.assign).}

\item{querytables}{Vector of names of the R symbols to assign in the Datashield R session on each server in \code{logins}.
The assigned R variables will be used as the input raw data.
Other assigned R variables in \code{funcPreProc} are ignored.}

\item{byColumn}{A logical value indicating whether the input data is centered by column or row.
Default, TRUE, centering by column. Constant variables across samples are removed. 
If FALSE, centering and scaling by row. Constant samples across variables are removed.}

\item{TOL}{Tolerance of 0}

\item{ind}{Index of table in querytables to compute SSCP. Default, ind = 1.}
}
\description{
Function for computing the federated SSCP matrix
}
\keyword{internal}
