% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{federateSNF}
\alias{federateSNF}
\title{Federated SNF}
\usage{
federateSNF(loginFD, logins, func, symbol, neighbors = 20, alpha = 0.5, iter = 20, TOL = 1e-10)
}
\arguments{
\item{loginFD}{Login information of the FD server}

\item{logins}{Login information of data repositories}

\item{func}{Encoded definition of a function for preparation of raw data matrices. 
Two arguments are required: conns (list of DSConnection-classes), 
symbol (name of the R symbol) (see datashield.assign).}

\item{symbol}{Encoded vector of names of the R symbols to assign in the Datashield R session on each server in \code{logins}.
The assigned R variables will be used as the input raw data.
Other assigned R variables in \code{func} are ignored.}

\item{TOL}{Tolerance of 0, deprecated}

\item{...}{see \code{SNFtool::SNF}}
}
\value{
SNF matrix
}
\description{
Function for SNF federated analysis on the virtual cohort combining multiple cohorts
}
