% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{solveSSCP}
\alias{solveSSCP}
\title{Find X from X*t(X) and t(X)*X}
\usage{
solveSSCP(XXt, XtX, r, Xr, TOL = 1e-10)
}
\arguments{
\item{XXt}{X*t(X)}

\item{XtX}{t(X)*X}

\item{r}{A non-null vector of length \code{ncol(t(X)*X)}}

\item{Xr}{A vector of length \code{nrow(X * t(X))}, equals to the product X %*% r}
}
\value{
X
}
\description{
Find X from X*t(X)  and t(X)*X
}
\keyword{internal}
